WIP: Add rspec-matcher `look_like_before`
I've been playing around with creating an rspec matcher that can help
you keep track of the visual appearance of your app. Something like

   expect(url).to look_like_before(width: 320)

The first time it's run, it will pass and create a baseline image. This
baseline image is then commited to the repository. Next time you run the
spec a new snapshot image is created and compared with the old baseline.
If it's different, the spec is failed and a diff image is created. If
you made an intended change, you can simply ditch the diff image and
replace the baseline.